Fix a long known issue that faultinject case could abort if the sql file is not small.
It seems that when the sql file is larger than ~4K, the case could ~ abort at the 4K offset.
The reason seems to be that some spawn external commands (i.e. \! in sql file)  (ssh?) could
swallow stdin which is expected for psql.

A simple fix is to add "</dev/null" for all external commands in our sql file.

Another fix seems to be that that pg_regress should use "-f $sql_file" instead of "< $sql_file".
    snprintf(psql_cmd, sizeof(psql_cmd),
	-            "%s " SYSTEMQUOTE "\"%s%spsql\" -X -a -q -d \"%s\" < \"%s\" > \"%s\" 2>&1" SYSTEMQUOTE,
	+            "%s " SYSTEMQUOTE "\"%s%spsql\" -X -a -q -d \"%s\" -f \"%s\" > \"%s\" 2>&1" SYSTEMQUOTE,

No we fix in our case at first so that we could tolerate any code of pg_regress.

Of course, faultinject2_python is added back to faultinject_python and thus is removed.