Fix memory related errors (#622)
Fix memory related errors

When using sql to select a table that has an array type column, the
container will be crashed. This is because the code that handles sql
freed the memory for the result tuples before the server sends the
result to clients.

There is an example output of crashed container:
```
WARNING: forcibly aborting a subtransaction that has not been exited
CONTEXT: PLContainer function "memtest7"
ERROR: invalid memory alloc request size 18446744073709551613 (mcxt.c:1270)
CONTEXT: PLContainer function "memtest7"
```
The solution does not free the SPI_tuple manually, instead, let
postgre frees it automatically, or, you can free it after the data send
to clients.