Fix test cases for r 4.2+ (#680)
* fix syntax and functions for R 4.2+

Vector with more than one element cannot be used in logical expression,
need use |. According to:
Using vectors of more than one element in ‘&&’ or ‘||’ will give a
warning (as from R 4.2.0), or an error if the environment variable
‘R_CHECK_LENGTH_1_LOGIC2’ is set to a true value (this is
intended to become the default in future).

'if' cannot handle with vectors, need use 'ifelse', or use identical to
compare vectors in 'if'.

as.character's behavior changed in R 4.3(or it's a bug), use format.