Support bulk transmit >= 1 GB data to container (#694)
## Description

Currently, array_agg() is the primary way to transmit more than one
row of data from QD/QE to the container. That is, data is first
aggregated as an array and then the array is transmitted as a whole
to the container. This is much more efficient than transmitting data
row-by-row.

However, the drawback of the approach is that the aggregated array
cannot be larger than 1 GB. This limitation is cause by how the array
is represented in PostgreSQL and Greenplum.

This patch fixes the issue by implementing a table function
`plcontainer_apply()`. The function reads data from the query in its
first parameter. The data size can be > 1 GB and multiple rows can
be pulled and transmitted as a whole for better performance.

One limitation of `plcontainer_apply()` is that the function is
requied to take exactly 1 parameter, whose type is array of some
composite type, and the return type is required to be exactly
`SETOF record`. This is because PL/Container currently only supports
such types as parameters and results for bulk transmission. If we
want to support more data types, we can extend PL/Container to do so.

## Tests

See tests/regress/sql/apply.sql

## Additional Notes

## User Story or Issue