Fix plpy not in sys.modules
Per Python's doc, all loaded modules should be present in
`sys.modules`. However, this is not true for plpy because it is
mistakenly added to the table of built-in modules after init-ing the
interpreter.

This patch fixes the issue by moving the call to
`PyImport_AppendInittab()` in front of the call to `Py_Initialize()`
to confront to the doc, so that the plpy module can be added properly.