Support using process as backend for Python 3 (#713)
Using a process on host as backend to run the client makes debugging
much easier. This is expected to be done by 
`SET plcontainer.backend_type='process';` However, in the prevous
releases, this only works for Python 2 and R but not Python 3.

This patch fixes this issue by adding Python 3 and a valid client,
and fix the wrong UDS path issue for the backend type "process".
The issue was introduced by commit 699da5c.