Start client manually for process backend (#715)
Previously the backend was started automatically even when the
backend is a process. This makes it less convenient to monitor the
client. For example, sometimes `fork()` or `exec()` might fail and
finding the root cause.

This patch mitigate this issue by letting the user to start the
client manually. This gives user much more control on the process.
And the code can be largely simplified.