Only count running and exited containers in tests
Previously in test fault_inject_python.sql, we used `docker ps -a`
to count all containers regardless of the status. This will include
containers created but not used due to creation timeout.

It might be desirable to also clean those containers. But this is
not easy due to the existance of the cleaner process, because the
cleaner does not know whether a created cleaner will be used later.

Therefore, this patch fixes the issue by modifying only the test
to could only containers in "running" or "exited" status so that
the test result will not be affected by creation timeouts.